requirements.txt
Flask
Flask-Login
Flask-SQLAlchemy
psycopg2-binary
python-dotenv
Werkzeug
requests
pandas
stravalib
google-auth
google-auth-oauthlib
google-api-python-client
google-generativeai
openpyxl
markdown2
itsdangerous
gunicorn==21.2.0
